Sugested changes made on both Flight and Leg entities
