+++ src/main/java/acme/entities/leg/Leg.java
@@ -1,24 +1,27 @@
 
 package acme.entities.leg;
 
-import java.beans.Transient;
 import java.util.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+import javax.persistence.Transient;
 import javax.validation.Valid;
 
 import acme.client.components.basis.AbstractEntity;
+import acme.client.components.mappings.Automapped;
 import acme.client.components.validation.Mandatory;
-import acme.client.helpers.SpringHelper;
+import acme.client.components.validation.ValidMoment;
+import acme.client.components.validation.ValidString;
 import acme.constraints.ValidFlightNumber;
 import acme.entities.aircraft.Aircraft;
 import acme.entities.airline.Airline;
 import acme.entities.airport.Airport;
 import acme.entities.flight.Flight;
-import acme.entities.flight.FlightRepository;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -31,43 +34,41 @@ public class Leg extends AbstractEntity {
 
 	// Serialisation version --------------------------------------------------
 
-	private static final long serialVersionUID = 1L;
-
-	// Enumerate -------------------------------------------------------------
-
-
-	public enum LegStatus {
-
-		ON_TIME, DELAYED, CANCELLED, LANDED
-	}
+	private static final long	serialVersionUID	= 1L;
 
 	// Attributes -------------------------------------------------------------
 
-
-	// Queda abierto a posible Custom Constraint
-	// Se utiliza un validador ValidFlightNumber para comprobar que las primeras letras coinciden con las de la aerolinea
-	// Añadir una relación en el rol de AirlineManager @ManyToOne con Airline
 	@Mandatory
 	@Column(unique = true)
-	private String		flightNumber;
+	private String				flightNumber;
 
 	@Mandatory
-	private LegStatus	status;
+	@Valid
+	@Automapped
+	private LegStatus			status;
+
+	@Mandatory
+	@ValidMoment
+	@Temporal(TemporalType.TIMESTAMP)
+	private Date				scheduledDeparture;
+
+	@Mandatory
+	@ValidMoment
+	@Temporal(TemporalType.TIMESTAMP)
+	private Date				scheduledArrival;
+
+	@Mandatory
+	@ValidString(max = 50)
+	@Transient
+	private String				originCity;
+
+	@Mandatory
+	@ValidString(max = 50)
+	@Transient
+	private String				destinationCity;
 
 	// Derived attributes -----------------------------------------------------
 
-
-	@Transient
-	public Date getScheduledDeparture() {
-		FlightRepository repository = SpringHelper.getBean(FlightRepository.class);
-		return repository.findScheduledDeparture(this.getId());
-	};
-
-	@Transient
-	public Date getScheduledArrival() {
-		FlightRepository repository = SpringHelper.getBean(FlightRepository.class);
-		return repository.findScheduledArrival(this.getId());
-	};
 
 	@Transient
 	public Double getDuration() {
@@ -79,7 +80,7 @@ public class Leg extends AbstractEntity {
 
 		long durationInMilliseconds = this.getScheduledArrival().getTime() - this.getScheduledDeparture().getTime();
 		return durationInMilliseconds / (1000.0 * 60 * 60); // Convertir milisegundos a horas
-	};
+	}
 
 	// Relationships ----------------------------------------------------------
