+++ src/main/java/acme/entities/leg/LegStatus.java
@@ -1,0 +1,7 @@
+
+package acme.entities.leg;
+
+public enum LegStatus {
+
+	ON_TIME, DELAYED, CANCELLED, LANDED
+}
